Almost There!
Murder is clean and working, introductions good, intro and menu both good, platformer needs dialogue items and text added, but should be easy enough. Teagan is working on the "fight" scene between Act 2 and 3. About to copy murder scene and make necessary tweaks for the heist scene. Pushing as backup.