Added Part 3 Click Through
Added scene for the comic panel click-through before the platformer, however, it is using the storyboarded images. Waiting for the final comics to be completed by Alex so that they may be added. This push is also to act as a backup before syncing with the "fight scene" framework made by Teagan.